+++ npm/webpack-batteries-included-preprocessor/test/unit/index.spec.js
@@ -1,0 +1,22 @@
+const { expect } = require('chai')
+
+const preprocessor = require('../../index')
+
+describe('webpack-batteries-included-preprocessor', () => {
+  context('#getFullWebpackOptions', () => {
+    it('returns default webpack options (and does not add typescript config if no path specified)', () => {
+      const result = preprocessor.getFullWebpackOptions()
+
+      expect(result.node.global).to.be.true
+      expect(result.module.rules).to.have.length(3)
+      expect(result.resolve.extensions).to.eql(['.js', '.json', '.jsx', '.mjs', '.coffee'])
+    })
+
+    it('adds typescript config if path is specified', () => {
+      const result = preprocessor.getFullWebpackOptions('file/path', 'typescript/path')
+
+      expect(result.module.rules).to.have.length(4)
+      expect(result.module.rules[3].use[0].loader).to.include('ts-loader')
+    })
+  })
+})