+++ npm/webpack-dev-server/src/helpers/nextHandler.ts
@@ -29,7 +29,7 @@ export async function nextHandler (devServerConfig: WebpackDevServerConfig): Pro
  */
 function getNextJsPackages (devServerConfig: WebpackDevServerConfig) {
   const resolvePaths = { paths: [devServerConfig.cypressConfig.projectRoot] }
-  const packages = {} as { loadConfig: Function, getNextJsBaseWebpackConfig: Function }
+  const packages = {} as { loadConfig: Function, getNextJsBaseWebpackConfig: Function, nextLoadJsConfig: Function }
 
   try {
     const loadConfigPath = require.resolve('next/dist/server/config', resolvePaths)
@@ -45,6 +45,14 @@ function getNextJsPackages (devServerConfig: WebpackDevServerConfig) {
     packages.getNextJsBaseWebpackConfig = require(getNextJsBaseWebpackConfigPath).default
   } catch (e: any) {
     throw new Error(`Failed to load "next/dist/build/webpack-config" with error: ${ e.message ?? e}`)
+  }
+
+  try {
+    const loadJsConfigPath = require.resolve('next/dist/build/load-jsconfig', resolvePaths)
+
+    packages.nextLoadJsConfig = require(loadJsConfigPath).default
+  } catch (e: any) {
+    throw new Error(`Failed to load "next/dist/build/load-jsconfig" with error: ${ e.message ?? e}`)
   }
 
   return packages
@@ -131,13 +139,46 @@ function getNextJsPackages (devServerConfig: WebpackDevServerConfig) {
       middlewareMatchers?: MiddlewareMatcher[]
     }
   ]
+
+ * - v13.2.1
+  [
+    dir: string,
+    options:  {
+    buildId: string
+    config: NextConfigComplete
+    compilerType: CompilerNameValues
+    dev?: boolean
+    entrypoints: webpack.EntryObject
+    isDevFallback?: boolean
+    pagesDir?: string
+    reactProductionProfiling?: boolean
+    rewrites: CustomRoutes['rewrites']
+    runWebpackSpan: Span
+    target?: string
+    appDir?: string
+    middlewareMatchers?: MiddlewareMatcher[]
+    noMangling?: boolean
+    jsConfig: any
+    resolvedBaseUrl: string | undefined
+    supportedBrowsers: string[] | undefined
+    clientRouterFilters?: {
+        staticFilter: ReturnType<
+          import('../shared/lib/bloom-filter').BloomFilter['export']
+        >
+        dynamicFilter: ReturnType<
+          import('../shared/lib/bloom-filter').BloomFilter['export']
+        >
+      }
+    }
+  ]
  */
 async function loadWebpackConfig (devServerConfig: WebpackDevServerConfig): Promise<Configuration> {
-  const { loadConfig, getNextJsBaseWebpackConfig } = getNextJsPackages(devServerConfig)
+  const { loadConfig, getNextJsBaseWebpackConfig, nextLoadJsConfig } = getNextJsPackages(devServerConfig)
 
   const nextConfig = await loadConfig('development', devServerConfig.cypressConfig.projectRoot)
   const runWebpackSpan = getRunWebpackSpan(devServerConfig)
   const reactVersion = getReactVersion(devServerConfig.cypressConfig.projectRoot)
+  const jsConfigResult = await nextLoadJsConfig?.(devServerConfig.cypressConfig.projectRoot, nextConfig)
 
   const webpackConfig = await getNextJsBaseWebpackConfig(
     devServerConfig.cypressConfig.projectRoot,
@@ -155,6 +196,8 @@ async function loadWebpackConfig (devServerConfig: WebpackDevServerConfig): Prom
       compilerType: 'client',
       // Required for Next.js > 13
       hasReactRoot: reactVersion === 18,
+      // Required for Next.js > 13.2.1 to respect TS/JS config
+      jsConfig: jsConfigResult.jsConfig,
     },
   )