+++ packages/graphql/src/schemaTypes/inputTypes/gql-WizardUpdateInput.ts
@@ -1,12 +1,10 @@
 import { inputObjectType } from 'nexus'
-import { FrontendFrameworkEnum, SupportedBundlerEnum } from '../enumTypes/gql-WizardEnums'
+import { SupportedBundlerEnum } from '../enumTypes/gql-WizardEnums'
 
 export const WizardUpdateInput = inputObjectType({
   name: 'WizardUpdateInput',
   definition (t) {
-    t.field('framework', {
-      type: FrontendFrameworkEnum,
-    })
+    t.string('framework')
 
     t.field('bundler', {
       type: SupportedBundlerEnum,