+++ packages/launchpad/cypress/e2e/scaffold-component-testing.cy.ts
@@ -83,6 +83,26 @@ describe('scaffolding component testing', {
       cy.findByRole('button', { name: 'Continue' }).click()
       verifyConfigFile(`cypress.config.ts`)
     })
+
+    it('detects react dependency even if `package.json` is not declared in `exports`', () => {
+      cy.scaffoldProject('react-vite-ts-unconfigured')
+      cy.openProject('react-vite-ts-unconfigured')
+      cy.visitLaunchpad()
+      cy.skipWelcome()
+
+      cy.withCtx(async (ctx) => {
+        const reactPackageFilePath = 'node_modules/react/package.json'
+        const packageFileContent = await ctx.actions.file.readFileInProject(reactPackageFilePath)
+        const newPackageFileContents = packageFileContent.replace('"./package.json": "./package.json",', '')
+
+        await ctx.actions.file.writeFileInProject(reactPackageFilePath, newPackageFileContents)
+      })
+
+      cy.contains('Component Testing').click()
+      cy.get(`[data-testid="select-framework"]`)
+
+      cy.get('button').should('be.visible').contains('React.js(detected)')
+    })
   })
 
   context('vue3-vite-ts-unconfigured', () => {
@@ -147,6 +167,70 @@ describe('scaffolding component testing', {
       cy.contains('button', 'Next step').click()
       cy.findByRole('button', { name: 'Continue' }).click()
       verifyConfigFile(`cypress.config.js`)
+    })
+  })
+
+  context('3rd party ct plugin', () => {
+    it('Scaffolds component testing for Qwik using Vite', () => {
+      cy.scaffoldProject('qwik-app')
+      cy.openProject('qwik-app')
+
+      cy.withCtx(async (ctx) => {
+        await ctx.actions.file.removeFileInProject('./node_modules/cypress-ct-qwik')
+        await ctx.actions.file.moveFileInProject('./cypress-ct-qwik', './node_modules/cypress-ct-qwik')
+      })
+
+      cy.visitLaunchpad()
+      cy.skipWelcome()
+
+      cy.contains('Component Testing').click()
+      cy.contains('button', 'Qwik').should('be.visible')
+      cy.contains('button', 'Next step').click()
+
+      cy.findByTestId('dependencies-to-install').within(() => {
+        cy.contains('li', '@builder.io/qwik').within(() => {
+          cy.findByLabelText('installed')
+        })
+
+        cy.contains('li', 'vite').within(() => {
+          cy.findByLabelText('installed')
+        })
+      })
+
+      cy.contains('button', 'Continue').click()
+
+      verifyConfigFile('cypress.config.js')
+    })
+
+    it('Scaffolds component testing for Solid using Vite', () => {
+      cy.scaffoldProject('ct-public-api-solid-js')
+      cy.openProject('ct-public-api-solid-js')
+
+      cy.withCtx(async (ctx) => {
+        await ctx.actions.file.removeFileInProject('./node_modules/cypress-ct-solid-js')
+        await ctx.actions.file.moveFileInProject('./cypress-ct-solid-js', './node_modules/cypress-ct-solid-js')
+      })
+
+      cy.visitLaunchpad()
+      cy.skipWelcome()
+
+      cy.contains('Component Testing').click()
+      cy.contains('button', 'Solid').should('be.visible')
+      cy.contains('button', 'Next step').click()
+
+      cy.findByTestId('dependencies-to-install').within(() => {
+        cy.contains('li', 'solid-js').within(() => {
+          cy.findByLabelText('installed')
+        })
+
+        cy.contains('li', 'vite').within(() => {
+          cy.findByLabelText('installed')
+        })
+      })
+
+      cy.contains('button', 'Continue').click()
+
+      verifyConfigFile('cypress.config.js')
     })
   })
 })