+++ packages/server/lib/socket-base.ts
@@ -20,7 +20,7 @@ import { openFile, OpenFileDetails } from './util/file-opener'
 import open from './util/open'
 import type { DestroyableHttpServer } from './util/server_destroy'
 import * as session from './session'
-import { AutomationCookie, cookieJar, SameSiteContext, automationCookieToToughCookie } from './util/cookies'
+import { cookieJar, SameSiteContext, automationCookieToToughCookie, SerializableAutomationCookie } from './util/cookies'
 import runEvents from './plugins/run_events'
 
 // eslint-disable-next-line no-duplicate-imports
@@ -28,31 +28,11 @@ import type { Socket } from '@packages/socket'
 
 import type { RunState, CachedTestState } from '@packages/types'
 import { cors } from '@packages/network'
+import memory from './browsers/memory'
 
 type StartListeningCallbacks = {
   onSocketConnection: (socket: any) => void
 }
-
-const runnerEvents = [
-  'reporter:restart:test:run',
-  'runnables:ready',
-  'run:start',
-  'test:before:run:async',
-  'reporter:log:add',
-  'reporter:log:state:changed',
-  'paused',
-  'test:after:hooks',
-  'run:end',
-] as const
-
-const reporterEvents = [
-  // "go:to:file"
-  'runner:restart',
-  'runner:abort',
-  'runner:show:snapshot',
-  'runner:hide:snapshot',
-  'reporter:restarted',
-] as const
 
 const debug = Debug('cypress:server:socket-base')
 
@@ -389,7 +369,7 @@ export class SocketBase {
         })
       })
 
-      const setCrossOriginCookie = ({ cookie, url, sameSiteContext }: { cookie: AutomationCookie, url: string, sameSiteContext: SameSiteContext }) => {
+      const setCrossOriginCookie = ({ cookie, url, sameSiteContext }: { cookie: SerializableAutomationCookie, url: string, sameSiteContext: SameSiteContext }) => {
         const domain = cors.getOrigin(url)
 
         cookieJar.setCookie(automationCookieToToughCookie(cookie, domain), url, sameSiteContext)
@@ -466,6 +446,12 @@ export class SocketBase {
               return setCrossOriginCookie(args[0])
             case 'request:sent:with:credentials':
               return this.localBus.emit('request:sent:with:credentials', args[0])
+            case 'start:memory:profiling':
+              return memory.startProfiling(automation, args[0])
+            case 'end:memory:profiling':
+              return memory.endProfiling()
+            case 'check:memory:pressure':
+              return memory.checkMemoryPressure({ ...args[0], automation })
             default:
               throw new Error(`You requested a backend event we cannot handle: ${eventName}`)
           }
@@ -553,18 +539,6 @@ export class SocketBase {
           })
         })
       }
-
-      reporterEvents.forEach((event) => {
-        socket.on(event, (data) => {
-          this.toRunner(event, data)
-        })
-      })
-
-      runnerEvents.forEach((event) => {
-        socket.on(event, (data) => {
-          this.toReporter(event, data)
-        })
-      })
 
       callbacks.onSocketConnection(socket)