+++ scripts/wait-on-circle-jobs.js
@@ -15,10 +15,6 @@ const jobName = process.env.CIRCLE_JOB || 'wait-on-circle-jobs'
 
 const workflowId = process.env.CIRCLE_WORKFLOW_ID
 
-const branchesToAlwaysFinalize = ['develop']
-
-const requireAllJobsToPass = !branchesToAlwaysFinalize.includes(process.env.CIRCLE_BRANCH)
-
 const getAuth = () => `${process.env.CIRCLE_TOKEN}:`
 
 const verifyCI = () => {
@@ -86,7 +82,7 @@ const waitForAllJobs = async (jobNames, workflowId) => {
   const runningJobNames = _.map(runningJobs, 'name')
   const failedJobNames = _.map(failedJobs, 'name')
 
-  if (requireAllJobsToPass && _.intersection(jobNames, failedJobNames).length) {
+  if (_.intersection(jobNames, failedJobNames).length) {
     console.error('At least one failing job has prevented percy-finalize from running', failedJobs)
     process.exit(1)
   }
@@ -106,7 +102,7 @@ const waitForAllJobs = async (jobNames, workflowId) => {
   const jobsToWaitFor = _.intersection(jobNames, futureOrRunning)
 
   // logging something every time this runs will avoid CI timing out if there is no activity for 10 mins.
-  console.log(`waiting for jobs, jobs outstanding: ${response.items.length}`)
+  console.log(`waiting for jobs, jobs outstanding: ${jobsToWaitFor.length}`)
 
   debug('jobs to wait for %o', jobsToWaitFor)